Fixed typos

+++ smartfox/recipes/install_client.rb
@@ -13,6 +13,7 @@ bash "untar_smartfox" do
    tar zxvf "#{Chef::Config[:file_cache_path]}/smartfox.tar.gz" -C /usr/local
    ln -s /usr/local/SFS_PRO_1.6.6 /usr/local/smartfox
  EOM
+ not_if { ::File.exists?("/usr/local/SFS_PRO_1.6.6") }
 end
 
 template "/usr/local/smartfox/Server/config.xml" do
@@ -28,7 +29,7 @@ end
 
 bash "start_smartfox" do
   code <<-EOM
-    /usr/local/smartfox/sfsls start
+    /usr/local/smartfox/Server/sfsls start
   EOM
 end
